# Conflicts:
#	spring/src/main/java/com/givoo/repository/Users/UsersRepository.java
#	spring/src/main/java/com/givoo/service/DonationService.java
#	spring/src/main/java/com/givoo/service/OrganizationService.java
#	spring/src/main/java/com/givoo/service/UsersService.java
#	spring/src/main/java/com/givoo/service/serviceImp/DonationServiceImpl.java
#	spring/src/main/java/com/givoo/service/serviceImp/OrganizationServiceImpl.java
#	spring/src/main/java/com/givoo/service/serviceImp/UsersServiceImpl.java

+++ spring/src/main/java/com/givoo/entity/Users.java
@@ -12,7 +12,7 @@ public class Users {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "user_id")
-    private int userId;
+    private Long userId;
 
     @Column(name = "user_name", nullable = false)
     private String userName;
@@ -37,7 +37,7 @@ public class Users {
 
     @Column(name = "token", nullable = false)
     private String token;
-  
+
     @Column(name = "user_image")
     private String userImage;
